infrastructure/requirements.txt
aws-cdk-lib==2.115.0
constructs>=10.0.0
pytest==7.4.3
